Extract monitoring and serving code out of server.ts
Give them their own files; inject the Env container to each one.
